backend/pyproject.toml
[project]
name = "backend_OneFit"
version = "0.1.0"
description = "Le backend de l'application OneFit"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
